Rename long test names to be a bit shorter
Helps to avoid hitting path limits on Windows
