+++ src/test/run-pass-fulldeps/auxiliary/issue-16723.rs
