+++ src/test/run-pass/specialization/auxiliary/cross_crates_defaults.rs
